REF Move Keras frontends out of root module
Instead, place these in the respective frontend directories of 1D and
2D.
